backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-multipart==0.0.18
python-dotenv==1.0.1
pypdf==5.1.0
requests==2.32.5
